requirements.txt
fastapi
uvicorn
python-dotenv
pillow
google-generativeai
firebase-admin
stripe
python-multipart
